# controllers/contact.controllers.go
package controllers

import (
	"context"
	"database/sql"
	"net/http"
	"strconv"
	"time"

	db "golang-crudsqlc-rest/db/sqlc"
	"golang-crudsqlc-rest/schemas"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
)

type ContactController struct {
	db  *db.Queries
	ctx context.Context
}

func NewContactController(db *db.Queries, ctx context.Context) *ContactController {
	return &ContactController{db, ctx}
}

// Create contact  handler
//
//	@Summary	Create a contact
//	@Tags		contacts
//	@Accept		json
//	@Produce	json
//	@Param		payload	body		schemas.CreateContact	true	"Contact to create"
//	@Success	200		{object}	schemas.ContactResponse
//	@Failure	400		{object}	schemas.ErrorResponse
//	@Failure	502		{object}	schemas.ErrorResponse
//	@Router		/contacts/ [post]
func (cc *ContactController) CreateContact(ctx *gin.Context) {
	var payload *schemas.CreateContact

	if err := ctx.ShouldBindJSON(&payload); err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"status": "Failed payload", "error": err.Error()})
		return
	}

	now := time.Now()
	args := &db.CreateContactParams{
		FirstName:   payload.FirstName,
		LastName:    payload.LastName,
		PhoneNumber: payload.PhoneNumber,
		Street:      payload.Street,
		CreatedAt:   now,
		UpdatedAt:   now,
	}

	contact, err := cc.db.CreateContact(ctx, *args)

	if err != nil {
		ctx.JSON(http.StatusBadGateway, gin.H{"status": "Failed retrieving contact", "error": err.Error()})
		return
	}

	ctx.JSON(http.StatusOK, gin.H{"status": "successfully created contact", "contact": contact})
}

// Update contact handler
//
//	@Summary	Update a contact
//	@Tags		contacts
//	@Accept		json
//	@Produce	json
//	@Param		contactId	path		string					true	"Contact ID"	format(uuid)
//	@Param		payload		body		schemas.UpdateContact	true	"Fields to update"
//	@Success	200			{object}	schemas.ContactResponse
//	@Failure	400			{object}	schemas.ErrorResponse
//	@Failure	404			{object}	schemas.ErrorResponse
//	@Failure	502			{object}	schemas.ErrorResponse
//	@Router		/contacts/{contactId} [patch]
func (cc *ContactController) UpdateContact(ctx *gin.Context) {
	var payload *schemas.UpdateContact
	contactId := ctx.Param("contactId")

	if err := ctx.ShouldBindJSON(&payload); err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"status": "Failed payload", "error": err.Error()})
		return
	}

	now := time.Now()
	args := &db.UpdateContactParams{
		ContactID:   uuid.MustParse(contactId),
		FirstName:   sql.NullString{String: payload.FirstName, Valid: payload.FirstName != ""},
		LastName:    sql.NullString{String: payload.LastName, Valid: payload.LastName != ""},
		PhoneNumber: sql.NullString{String: payload.PhoneNumber, Valid: payload.PhoneNumber != ""},
		Street:      sql.NullString{String: payload.PhoneNumber, Valid: payload.Street != ""},
		UpdatedAt:   sql.NullTime{Time: now, Valid: true},
	}

	contact, err := cc.db.UpdateContact(ctx, *args)

	if err != nil {
		if err == sql.ErrNoRows {
			ctx.JSON(http.StatusNotFound, gin.H{"status": "failed", "message": "Failed to retrieve contact with this ID"})
			return
		}
		ctx.JSON(http.StatusBadGateway, gin.H{"status": "Failed retrieving contact", "error": err.Error()})
		return
	}

	ctx.JSON(http.StatusOK, gin.H{"status": "successfully updated contact", "contact": contact})
}

// Get a single handler
//
//	@Summary	Get a contact by ID
//	@Tags		contacts
//	@Produce	json
//	@Param		contactId	path		string	true	"Contact ID"	format(uuid)
//	@Success	200			{object}	schemas.ContactResponse
//	@Failure	404			{object}	schemas.ErrorResponse
//	@Failure	502			{object}	schemas.ErrorResponse
//	@Router		/contacts/{contactId} [get]
func (cc *ContactController) GetContactById(ctx *gin.Context) {
	contactId := ctx.Param("contactId")

	contact, err := cc.db.GetContactById(ctx, uuid.MustParse(contactId))
	if err != nil {
		if err == sql.ErrNoRows {
			ctx.JSON(http.StatusNotFound, gin.H{"status": "failed", "message": "Failed to retrieve contact with this ID"})
			return
		}
		ctx.JSON(http.StatusBadGateway, gin.H{"status": "Failed retrieving contact", "error": err.Error()})
		return
	}

	ctx.JSON(http.StatusOK, gin.H{"status": "Successfully retrived id", "contact": contact})
}

// Retrieve all records handlers
//
//	@Summary	List contacts
//	@Tags		contacts
//	@Produce	json
//	@Param		page	query		int	false	"Page number"		default(1)
//	@Param		limit	query		int	false	"Contacts per page"	default(10)
//	@Success	200		{object}	schemas.ContactListResponse
//	@Failure	502		{object}	schemas.ErrorResponse
//	@Router		/contacts/ [get]
func (cc *ContactController) GetAllContacts(ctx *gin.Context) {
	var page = ctx.DefaultQuery("page", "1")
	var limit = ctx.DefaultQuery("limit", "10")

	reqPageID, _ := strconv.Atoi(page)
	reqLimit, _ := strconv.Atoi(limit)
	offset := (reqPageID - 1) * reqLimit

	args := &db.ListContactsParams{
		Limit:  int32(reqLimit),
		Offset: int32(offset),
	}

	contacts, err := cc.db.ListContacts(ctx, *args)
	if err != nil {
		ctx.JSON(http.StatusBadGateway, gin.H{"status": "Failed to retrieve contacts", "error": err.Error()})
		return
	}

	if contacts == nil {
		contacts = []db.Contact{}
	}

	ctx.JSON(http.StatusOK, gin.H{"status": "Successfully retrieved all contacts", "size": len(contacts), "contacts": contacts})
}

// Deleting contact handlers
//
//	@Summary	Delete a contact
//	@Tags		contacts
//	@Produce	json
//	@Param		contactId	path	string	true	"Contact ID"	format(uuid)
//	@Success	204
//	@Failure	404	{object}	schemas.ErrorResponse
//	@Failure	502	{object}	schemas.ErrorResponse
//	@Router		/contacts/{contactId} [delete]
func (cc *ContactController) DeleteContactById(ctx *gin.Context) {
	contactId := ctx.Param("contactId")

	_, err := cc.db.GetContactById(ctx, uuid.MustParse(contactId))
	if err != nil {
		if err == sql.ErrNoRows {
			ctx.JSON(http.StatusNotFound, gin.H{"status": "failed", "message": "Failed to retrieve contact with this ID"})
			return
		}
		ctx.JSON(http.StatusBadGateway, gin.H{"status": "Failed retrieving contact", "error": err.Error()})
		return
	}

	err = cc.db.DeleteContact(ctx, uuid.MustParse(contactId))
	if err != nil {
		ctx.JSON(http.StatusBadGateway, gin.H{"status": "failed", "error": err.Error()})
		return
	}

	ctx.JSON(http.StatusNoContent, gin.H{"status": "successfuly deleted"})

}